move notification  interfaces
